# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Management API",
        "description": "API for managing school operations including students, teachers, courses, grades, and attendance",
        "contact": {
            "email": "admin@school.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/api/estudiantes": {
            "get": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Obtiene la lista de estudiantes con opciones de filtro y visualización.",
                "operationId": "203351e8ece842d3a3513835b24ace25",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Filtrar por Nombre completo del estudiante.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de estudiantes obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre_completo": {
                                                "type": "string",
                                                "example": "Juan Perez"
                                            },
                                            "fecha_de_nacimiento": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2000-01-01"
                                            },
                                            "padre_madre_o_tutor": {
                                                "type": "string",
                                                "example": "Maria Lopez"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Crea un nuevo estudiante.",
                "operationId": "b4c3d28d31f5b43c5b25573e1a10a98a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "identification_number",
                                    "gender",
                                    "date_of_birth",
                                    "phone",
                                    "address",
                                    "parent_name",
                                    "parent_phone",
                                    "parent_email",
                                    "emergency_contact_name",
                                    "emergency_contact_phone",
                                    "emergency_contact_email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Perez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan.perez@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "identification_number": {
                                        "type": "string",
                                        "example": "123-456789-0"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female",
                                            "other"
                                        ],
                                        "example": "male"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2005-05-10"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "123-456-7890"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Calle Falsa 123"
                                    },
                                    "parent_name": {
                                        "type": "string",
                                        "example": "Maria Lopez"
                                    },
                                    "parent_phone": {
                                        "type": "string",
                                        "example": "987-654-3210"
                                    },
                                    "parent_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "maria.lopez@example.com"
                                    },
                                    "emergency_contact_name": {
                                        "type": "string",
                                        "example": "Pedro Lopez"
                                    },
                                    "emergency_contact_phone": {
                                        "type": "string",
                                        "example": "555-123-4567"
                                    },
                                    "emergency_contact_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "pedro.lopez@example.com"
                                    },
                                    "profile_image": {
                                        "description": "Optional: Student's profile image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "grade_level": {
                                        "type": "string",
                                        "example": "10th Grade"
                                    },
                                    "enrollment_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-09-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Estudiante creado exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Student created successfully"
                                        },
                                        "student": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre_completo": {
                                                    "type": "string",
                                                    "example": "Juan Perez"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}